Add todo for workaround
