+++ mars/tensor/core.py
@@ -190,19 +190,6 @@ class Tensor(Entity):
     __slots__ = ()
     _allow_data_type_ = (TensorData,)
 
-    def __dir__(self):
-        from ..lib.lib_utils import dir2
-        obj_dir = dir2(self)
-        if self._data is not None:
-            obj_dir = sorted(set(dir(self._data) + obj_dir))
-        return obj_dir
-
-    def __str__(self):
-        return self._data.__str__()
-
-    def __repr__(self):
-        return self._data.__repr__()
-
     def __len__(self):
         return len(self._data)
